backend/pyproject.toml
[project]
name = "personality-twin-backend"
version = "0.1.0"
description = "AI-powered personality analysis backend using FastAPI and scikit-learn"
authors = [{name = "TwinMind", email = "team@twinmind.ai"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "psycopg2-binary>=2.9.0",
]
